File restructure (#3)
* Remove sys::utils
* Extract _sys crate from _64
