+++ src/event/mod.rs
@@ -1,4 +1,4 @@
-use crate::sys::*;
+use _sys::*;
 
 pub struct EventFeed;
 pub enum Event {
