+++ src/render/buffer.rs
@@ -1,4 +1,4 @@
-use crate::sys::*;
+use _sys::*;
 
 pub struct Buffer {
     _type: GLenum,
