+++ src/render/vertex.rs
@@ -1,4 +1,4 @@
-use crate::sys::*;
+use _sys::*;
 use core::mem::size_of;
 
 pub trait Vertex: Sized {
